backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery[redis]
Pillow
python-multipart
pydantic
python-dotenv
pandas
numpy
minio
scikit-learn
icecream
pydantic-settings

torch
torchvision
torchaudio
ultralytics
easyocr
opencv-python-headless
PyYAML
transformers
pymorphy3
razdel

uvicorn[standard]
click